refactor(WPN-1652): extract SnapAssetsAdapter for Solana balance reads
