+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/snap-solana-wallet.git"
   },
   "source": {
-    "shasum": "plRX1XTzFRIxN8mAd+hRn7XmQH03DosKAy2PNVyo1B4=",
+    "shasum": "gimlsTQaSaGjPb9b20pZcxJV1Xy0+qZ8MLPBTyIiWAQ=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
