+++ packages/snap/src/core/services/assets/index.ts
@@ -1,3 +1,4 @@
+export * from './adapters/SnapAssetsAdapter';
 export * from './AssetsRepository';
 export * from './AssetsService';
 export * from './TokenHelper';
